Migrate the web app to Mantine 9.4
Bump @mantine/core, @mantine/dropzone, and @mantine/hooks from 9.1 to
9.4, the bump held back from the non-critical group because it changes
how a collapsed Collapse renders.

Since 9.4 a closed Collapse holds its content in a hidden React Activity
boundary rather than rendering it clipped-but-present, and no longer
unmounts the panel under a reduced-motion preference. The server renderer
omits hidden Activity content, so the InvitationTerms render assertions
that read collapsed detail from renderToStaticMarkup can no longer see
it. Move that terms-render block to a browser test that mounts the panel,
expands every disclosure, and asserts the rendered markup, beside the
existing InvitationTerms browser tests. Update the reduced-motion browser
test to assert the panel now stays mounted but hidden rather than
unmounting, and refresh the component comments that described the old
unmount behavior.
