+++ apps/web/test/unit/acceptConsent.test.ts
@@ -11,7 +11,6 @@ import {
 } from "@psilink/core";
 
 import { AcceptInvitationPanel } from "@components/AcceptInvitationPanel";
-import { InvitationTerms } from "@components/InvitationTerms";
 
 import {
   TRANSFORM_FUNCTION_GLOSSARY,
@@ -1377,403 +1376,6 @@ describe("summarizeInvitation", () => {
     expect(entry).toContain("mystery");
     expect(entry).not.toContain(BEL);
     expect(entry).toContain("\\x07");
-  });
-});
-
-describe("accept screen: terms render from a decoded token", () => {
-  test("renders the inviter's linkage terms for review", () => {
-    const html = renderPanel({
-      decode: { status: "ready", invitation: makeInvitation() },
-    });
-    expect(html).toContain("Invitation from County Health Department");
-    expect(html).toContain("SSN + LN + DOB");
-    expect(html).toContain("Social Security number");
-    // psi (not psi-c): the identifier-revealing description.
-    expect(html).toContain("shared identifiers");
-    expect(html).toContain("MOU-2025-0042");
-    expect(html).toContain("risk_score");
-    expect(html).toContain("program_outcome");
-  });
-
-  test("renders the deduplicate setting as a duplicate-matches note", () => {
-    const many = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({ deduplicate: true }),
-      },
-    });
-    expect(many).toContain("may match more than one");
-    const one = renderPanel({
-      decode: { status: "ready", invitation: makeInvitation() },
-    });
-    expect(one).toContain("matches at most one");
-  });
-
-  test("renders psi-c as the count-only description, flagged not-yet-applied", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({ algorithm: "psi-c" }),
-      },
-    });
-    expect(html).toContain("number of records");
-    // psi-c is a disclosure guarantee the run does not yet honor, so the
-    // count-only line is flagged proposed-but-not-applied -- the matched
-    // identifiers are still revealed. This note pairs with PSI_C gating in
-    // APPLIED_SETTINGS and fails loudly if that flag flips without updating the
-    // copy.
-    expect(html).toContain("does not yet apply it");
-    expect(html).toContain("still revealed");
-  });
-
-  test("renders the transform, swap, fuzzy, and constraint rules that affect matching", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [
-            {
-              name: "first_name",
-              type: "first_name",
-              constraints: { allowedCharacters: "A-Z " },
-            },
-            { name: "last_name", type: "last_name" },
-            { name: "dob", type: "date_of_birth" },
-          ],
-          linkageKeys: [
-            {
-              name: "FN1 + DOB",
-              elements: [
-                {
-                  field: "first_name",
-                  transform: [
-                    { function: "substring", params: { start: 1, length: 1 } },
-                  ],
-                },
-                { field: "dob", generateFuzzyComparisons: "adjacent_years" },
-              ],
-            },
-            {
-              name: "swap(LN, FN) + DOB",
-              elements: [
-                { field: "last_name" },
-                { field: "first_name" },
-                { field: "dob" },
-              ],
-              swap: ["last_name", "first_name"],
-            },
-          ],
-        }),
-      },
-    });
-    // The transform leads with its literal slice phrase, and the fuzzy expansion
-    // surfaces its plain-language label on the elements.
-    expect(html).toContain("Matches on the first character");
-    // ... including its parameters, each on its own line, never joined: a
-    // separator inside one value must not read as additional parameters
-    // (mirrors the payload-column disambiguation below).
-    expect(html).toContain("start: 1");
-    expect(html).toContain("length: 1");
-    expect(html).not.toContain("start: 1, length: 1");
-    expect(html).toContain("adjacent years");
-    // The swap is described in field-label terms.
-    expect(html).toContain(
-      "Last name and First name may be matched in either order",
-    );
-    // The field constraint surfaces under the data used.
-    expect(html).toContain("characters limited to A-Z");
-  });
-
-  test("renders a generic swap note when the swapped fields share a label", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [{ name: "first_name", type: "first_name" }],
-          linkageKeys: [
-            {
-              name: "alias swap",
-              elements: [
-                { field: "first_name", name: "given" },
-                { field: "first_name", name: "preferred" },
-              ],
-              swap: ["given", "preferred"],
-            },
-          ],
-        }),
-      },
-    });
-    // The swap is still surfaced, but the note is generic rather than naming
-    // "First name and First name".
-    expect(html).toContain(
-      "Two of these elements may be matched in either order",
-    );
-    expect(html).not.toContain("First name and First name");
-  });
-
-  test("shows no breadth marker for a plain term", () => {
-    const html = renderPanel({
-      decode: { status: "ready", invitation: makeInvitation() },
-    });
-    // baseTerms carries a single key with no transform, swap, or fuzzy rule, so
-    // its header one-liner names the fields with no breadth marker.
-    expect(html).toContain("Matches on SSN - last name - date of birth");
-    expect(html).not.toContain("(partial)");
-  });
-
-  test("marks only the elements that loosen matching in the header one-liner", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [
-            { name: "ssn", type: "ssn" },
-            { name: "dob", type: "date_of_birth" },
-          ],
-          linkageKeys: [
-            {
-              name: "plain key",
-              elements: [{ field: "ssn" }, { field: "dob" }],
-            },
-            {
-              name: "fuzzy key",
-              elements: [
-                { field: "dob", generateFuzzyComparisons: "adjacent_years" },
-              ],
-            },
-          ],
-        }),
-      },
-    });
-    // The fuzzy key's header carries the "(fuzzy)" marker; the plain key's header
-    // names its fields with none. Exactly one element loosens matching, so the
-    // header marker appears once.
-    expect(html).toContain("Matches on date of birth (fuzzy)");
-    expect(html).toContain("Matches on SSN - date of birth");
-    const markers = html.match(/\(fuzzy\)/g) ?? [];
-    expect(markers.length).toBe(1);
-  });
-
-  test("renders payload columns as separate items so a comma in a name cannot merge them", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          payload: {
-            send: [{ name: "risk, score" }, { name: "tier" }],
-            receive: [],
-          },
-        }),
-      },
-    });
-    // The comma-bearing name renders intact ...
-    expect(html).toContain("risk, score");
-    expect(html).toContain("tier");
-    // ... and as its own item: a join would have produced
-    // "risk, score, tier", making two columns read as three.
-    expect(html).not.toContain("score, tier");
-  });
-
-  test("escapes injection characters in the rendered identity", () => {
-    const html = render(
-      createElement(InvitationTerms, {
-        linkageTerms: { ...baseTerms, identity: EVIL_IDENTITY },
-      }),
-    );
-    expect(html).not.toContain(ESC);
-    expect(html).not.toContain(RLO);
-    expect(html).toContain("\\x1b");
-  });
-
-  test("renders the transform-parameter overflow marker", () => {
-    const params: Record<string, number> = {};
-    for (let i = 0; i < 20; i += 1) params["p" + i] = i;
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [{ name: "ssn", type: "ssn" }],
-          linkageKeys: [
-            {
-              name: "K",
-              elements: [
-                { field: "ssn", transform: [{ function: "f", params }] },
-              ],
-            },
-          ],
-        }),
-      },
-    });
-    // 16 parameters render, then the overflow marker for the remaining 4.
-    expect(html).toContain("... 4 more");
-  });
-
-  test("renders the runtime-coercion annotation for a coerced parameter", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [{ name: "ssn", type: "ssn" }],
-          linkageKeys: [
-            {
-              name: "K",
-              elements: [
-                {
-                  field: "ssn",
-                  transform: [
-                    {
-                      function: "replace_regex",
-                      params: { pattern: "x", replacement: null },
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        }),
-      },
-    });
-    // The declared null reaches the screen verbatim, and what the function
-    // actually runs is surfaced as a separate note, so the consent term cannot
-    // misstate the match yet the executed value is not folded into the
-    // partner-controlled param line.
-    expect(html).toContain("replacement: null");
-    expect(html).toContain("replacement runs as the empty string");
-    // The note carries a screen-reader-only "Runtime note:" lead-in marking it
-    // as system-authored -- a signal a partner (who controls only param-value
-    // text) cannot inject.
-    expect(html).toContain("Runtime note:");
-  });
-
-  test("a forged 'runs as' in a param value renders no system coercion note", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [{ name: "ssn", type: "ssn" }],
-          linkageKeys: [
-            {
-              name: "K",
-              elements: [
-                {
-                  field: "ssn",
-                  transform: [
-                    {
-                      function: "replace_regex",
-                      params: {
-                        pattern: "x",
-                        replacement: "Y runs as the empty string",
-                      },
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        }),
-      },
-    });
-    // The forged text renders as the verbatim param line ...
-    expect(html).toContain("replacement: Y runs as the empty string");
-    // ... but no genuine system note is synthesized: the "Runtime note:" lead-in
-    // a partner cannot inject is absent, so a screen reader does not hear this as
-    // a system-authored coercion note.
-    expect(html).not.toContain("Runtime note:");
-  });
-
-  test("flags a proposed deduplicate setting the current exchange does not apply", () => {
-    const proposed = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({ deduplicate: true }),
-      },
-    });
-    expect(proposed).toContain("may match more than one");
-    expect(proposed).toContain("does not yet apply it");
-    // The default (deduplicate off) matches the run, so it carries no flag.
-    const off = renderPanel({
-      decode: { status: "ready", invitation: makeInvitation() },
-    });
-    expect(off).toContain("matches at most one");
-    expect(off).not.toContain("does not yet apply it");
-  });
-
-  test("flags a proposed fuzzy comparison the current exchange does not apply", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [{ name: "dob", type: "date_of_birth" }],
-          linkageKeys: [
-            {
-              name: "DOB",
-              elements: [
-                { field: "dob", generateFuzzyComparisons: "adjacent_years" },
-              ],
-            },
-          ],
-        }),
-      },
-    });
-    expect(html).toContain("adjacent years");
-    expect(html).toContain("(proposed; not yet applied)");
-  });
-
-  test("depicts the transform interchange for a swap whose elements both transform", () => {
-    const html = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [
-            { name: "first_name", type: "first_name" },
-            { name: "last_name", type: "last_name" },
-          ],
-          linkageKeys: [
-            {
-              name: "Name",
-              elements: [
-                {
-                  field: "first_name",
-                  transform: [{ function: "to_upper_case" }],
-                },
-                {
-                  field: "last_name",
-                  transform: [{ function: "to_upper_case" }],
-                },
-              ],
-              swap: ["first_name", "last_name"],
-            },
-          ],
-        }),
-      },
-    });
-    // The generic swap note, plus the interchange detail naming which field each
-    // element's transforms run on once the fields swap.
-    expect(html).toContain("may be matched in either order");
-    expect(html).toContain("transforms shown for First name");
-    expect(html).toContain("are applied to Last name");
-
-    // A swap whose elements carry no transforms keeps the generic note alone.
-    const plain = renderPanel({
-      decode: {
-        status: "ready",
-        invitation: makeInvitation({
-          linkageFields: [
-            { name: "first_name", type: "first_name" },
-            { name: "last_name", type: "last_name" },
-          ],
-          linkageKeys: [
-            {
-              name: "Name",
-              elements: [{ field: "first_name" }, { field: "last_name" }],
-              swap: ["first_name", "last_name"],
-            },
-          ],
-        }),
-      },
-    });
-    expect(plain).toContain("may be matched in either order");
-    expect(plain).not.toContain("transforms shown for");
   });
 });
